python-service/requirements.txt
tensorflow-macos==2.15.0
tensorflow-metal==1.1.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
pillow==10.1.0
numpy==1.24.3
opencv-python==4.8.1.78
python-multipart==0.0.6
pydantic==2.5.0
scikit-learn==1.3.2
matplotlib==3.8.2
pandas==2.1.3
nltk==3.8.1